Refactoring the format module for excelx
